refactor(share): route View imports through widget barrels + use AssetClass type
Export *View presentational components through each widget's barrel index.ts
so kindPanelRegistry.tsx imports via slice barrels (@/widgets/overall, /news,
/fundamental, /financials, /options, /fear-greed) instead of deep internal
paths, per CLAUDE.md production-code import rule.

Also import AssetClass from @/shared/config/marketProfile in OverallView to
avoid inline union drift from the shared type.

Fix pre-existing vi.mock('react-markdown') factory in symbol-metadata.test.ts
to return { default: () => null } (vitest ESM default-export contract).
