+++ src/widgets/financials/index.ts
@@ -1,4 +1,7 @@
-export { FinancialsAiSummary } from './FinancialsAiSummary';
+export {
+    FinancialsAiSummary,
+    FinancialsAiSummaryView,
+} from './FinancialsAiSummary';
 export { FinancialsScorecard } from './FinancialsScorecard';
 export { FinancialsStatements } from './FinancialsStatements';
 export { PeriodToggle } from './PeriodToggle';
