+++ src/widgets/fundamental/index.ts
@@ -1,5 +1,8 @@
 // fundamental widget barrel — public API for external consumers.
 
-export { FundamentalAiSummary } from './FundamentalAiSummary';
+export {
+    FundamentalAiSummary,
+    FundamentalAiSummaryView,
+} from './FundamentalAiSummary';
 export { FundamentalAiSummaryError } from './FundamentalAiSummaryError';
 export { FundamentalAiSummarySkeleton } from './FundamentalAiSummarySkeleton';
