+++ src/widgets/news/index.ts
@@ -1,6 +1,6 @@
 // news widget barrel — public API for external consumers.
 
-export { NewsAiSummary } from './NewsAiSummary';
+export { NewsAiSummary, NewsAiSummaryView } from './NewsAiSummary';
 export { NewsAiSummaryError } from './NewsAiSummaryError';
 export { NewsAiSummaryErrorBoundary } from './NewsAiSummaryErrorBoundary';
 export { NewsAiSummarySkeleton } from './NewsAiSummarySkeleton';
