+++ src/widgets/overall/index.ts
@@ -6,3 +6,5 @@ export { OverallFactsSummary } from './OverallFactsSummary';
 export { OverallTriggerCta } from './OverallTriggerCta';
 export { ReanalyzeButton } from './ReanalyzeButton';
 export { DependencyProgress } from './DependencyProgress';
+// 'use client' presentational view — used by share panel registry.
+export { OverallView } from './OverallView';
