+++ src/widgets/share/ui/kindPanelRegistry.tsx
@@ -20,13 +20,13 @@ import type { Bar } from '@y0ngha/siglens-core';
 import { clusterKeyLevels, validateKeyLevels } from '@y0ngha/siglens-core';
 import { AnalysisPanel } from '@/widgets/analysis';
 import { ShareCandlestickChart } from '@/widgets/chart/ShareCandlestickChart';
-import { OverallView } from '@/widgets/overall/OverallView';
-import { NewsAiSummaryView } from '@/widgets/news/NewsAiSummary';
-import { FundamentalAiSummaryView } from '@/widgets/fundamental/FundamentalAiSummary';
-import { FinancialsAiSummaryView } from '@/widgets/financials/FinancialsAiSummary';
+import { OverallView } from '@/widgets/overall';
+import { NewsAiSummaryView } from '@/widgets/news';
+import { FundamentalAiSummaryView } from '@/widgets/fundamental';
+import { FinancialsAiSummaryView } from '@/widgets/financials';
 import { CongressTrendSummaryView } from '@/widgets/congress';
-import { OptionsAiAnalysisView } from '@/widgets/options/OptionsAiAnalysis';
-import { FearGreedShareView } from '@/widgets/fear-greed/FearGreedShareView';
+import { OptionsAiAnalysisView } from '@/widgets/options';
+import { FearGreedShareView } from '@/widgets/fear-greed';
 
 /**
  * AnalysisPanel has several required props that are interaction/live-data
